+++ src/ctl_externs.h
@@ -1,0 +1,324 @@
+#ifndef JEMALLOC_INTERNAL_CTL_EXTERNS_H
+#define JEMALLOC_INTERNAL_CTL_EXTERNS_H
+
+#include "jemalloc/internal/ctl.h"
+
+#define CTL_EXTERN(n)                                                          \
+	int n##_ctl(tsd_t *tsd, const size_t *mib, size_t miblen,              \
+	    void *oldp, size_t *oldlenp, void *newp, size_t newlen);
+
+CTL_EXTERN(approximate_stats_active)
+CTL_EXTERN(arena_i_decay)
+CTL_EXTERN(arena_i_destroy)
+CTL_EXTERN(arena_i_dirty_decay_ms)
+CTL_EXTERN(arena_i_dss)
+CTL_EXTERN(arena_i_extent_hooks)
+CTL_EXTERN(arena_i_initialized)
+CTL_EXTERN(arena_i_muzzy_decay_ms)
+CTL_EXTERN(arena_i_name)
+CTL_EXTERN(arena_i_oversize_threshold)
+CTL_EXTERN(arena_i_purge)
+CTL_EXTERN(arena_i_reset)
+CTL_EXTERN(arena_i_retain_grow_limit)
+CTL_EXTERN(arenas_bin_i_nregs)
+CTL_EXTERN(arenas_bin_i_nshards)
+CTL_EXTERN(arenas_bin_i_size)
+CTL_EXTERN(arenas_bin_i_slab_size)
+CTL_EXTERN(arenas_create)
+CTL_EXTERN(arenas_dirty_decay_ms)
+CTL_EXTERN(arenas_hugepage)
+CTL_EXTERN(arenas_lextent_i_size)
+CTL_EXTERN(arenas_lookup)
+CTL_EXTERN(arenas_muzzy_decay_ms)
+CTL_EXTERN(arenas_narenas)
+CTL_EXTERN(arenas_nbins)
+CTL_EXTERN(arenas_nhbins)
+CTL_EXTERN(arenas_nlextents)
+CTL_EXTERN(arenas_page)
+CTL_EXTERN(arenas_quantum)
+CTL_EXTERN(arenas_tcache_max)
+CTL_EXTERN(background_thread)
+CTL_EXTERN(config_cache_oblivious)
+CTL_EXTERN(config_debug)
+CTL_EXTERN(config_fill)
+CTL_EXTERN(config_infallible_new)
+CTL_EXTERN(config_lazy_lock)
+CTL_EXTERN(config_malloc_conf)
+CTL_EXTERN(config_opt_safety_checks)
+CTL_EXTERN(config_prof)
+CTL_EXTERN(config_prof_frameptr)
+CTL_EXTERN(config_prof_libgcc)
+CTL_EXTERN(config_prof_libunwind)
+CTL_EXTERN(config_stats)
+CTL_EXTERN(config_utrace)
+CTL_EXTERN(config_xmalloc)
+CTL_EXTERN(experimental_arenas_create_ext)
+CTL_EXTERN(experimental_hooks_prof_backtrace)
+CTL_EXTERN(experimental_hooks_prof_dump)
+CTL_EXTERN(experimental_hooks_prof_sample)
+CTL_EXTERN(experimental_hooks_prof_sample_free)
+CTL_EXTERN(experimental_hooks_thread_event)
+CTL_EXTERN(experimental_prof_recent_alloc_dump)
+CTL_EXTERN(experimental_prof_recent_alloc_max)
+CTL_EXTERN(experimental_utilization_batch_query)
+CTL_EXTERN(lg_prof_sample)
+CTL_EXTERN(max_background_threads)
+CTL_EXTERN(opt_abort)
+CTL_EXTERN(opt_abort_conf)
+CTL_EXTERN(opt_background_thread)
+CTL_EXTERN(opt_cache_oblivious)
+CTL_EXTERN(opt_confirm_conf)
+CTL_EXTERN(opt_debug_double_free_max_scan)
+CTL_EXTERN(opt_dirty_decay_ms)
+CTL_EXTERN(opt_disable_large_size_classes)
+CTL_EXTERN(opt_dss)
+CTL_EXTERN(opt_experimental_hpa_enforce_hugify)
+CTL_EXTERN(opt_experimental_hpa_max_purge_nhp)
+CTL_EXTERN(opt_experimental_hpa_start_huge_if_thp_always)
+CTL_EXTERN(opt_experimental_tcache_gc)
+CTL_EXTERN(opt_hpa)
+CTL_EXTERN(opt_hpa_dirty_mult)
+CTL_EXTERN(opt_hpa_hugification_threshold)
+CTL_EXTERN(opt_hpa_hugify_delay_ms)
+CTL_EXTERN(opt_hpa_hugify_style)
+CTL_EXTERN(opt_hpa_hugify_sync)
+CTL_EXTERN(opt_hpa_min_purge_delay_ms)
+CTL_EXTERN(opt_hpa_min_purge_interval_ms)
+CTL_EXTERN(opt_hpa_purge_threshold)
+CTL_EXTERN(opt_hpa_sec_max_alloc)
+CTL_EXTERN(opt_hpa_sec_max_bytes)
+CTL_EXTERN(opt_hpa_sec_nshards)
+CTL_EXTERN(opt_hpa_slab_max_alloc)
+CTL_EXTERN(opt_huge_arena_pac_thp)
+CTL_EXTERN(opt_junk)
+CTL_EXTERN(opt_lg_extent_max_active_fit)
+CTL_EXTERN(opt_lg_prof_interval)
+CTL_EXTERN(opt_lg_prof_sample)
+CTL_EXTERN(opt_lg_san_uaf_align)
+CTL_EXTERN(opt_lg_tcache_flush_large_div)
+CTL_EXTERN(opt_lg_tcache_flush_small_div)
+CTL_EXTERN(opt_lg_tcache_nslots_mul)
+CTL_EXTERN(opt_malloc_conf_env_var)
+CTL_EXTERN(opt_malloc_conf_global_var)
+CTL_EXTERN(opt_malloc_conf_symlink)
+CTL_EXTERN(opt_max_background_threads)
+CTL_EXTERN(opt_metadata_thp)
+CTL_EXTERN(opt_mutex_max_spin)
+CTL_EXTERN(opt_muzzy_decay_ms)
+CTL_EXTERN(opt_narenas)
+CTL_EXTERN(opt_oversize_threshold)
+CTL_EXTERN(opt_percpu_arena)
+CTL_EXTERN(opt_process_madvise_max_batch)
+CTL_EXTERN(opt_prof)
+CTL_EXTERN(opt_prof_accum)
+CTL_EXTERN(opt_prof_active)
+CTL_EXTERN(opt_prof_bt_max)
+CTL_EXTERN(opt_prof_final)
+CTL_EXTERN(opt_prof_gdump)
+CTL_EXTERN(opt_prof_leak)
+CTL_EXTERN(opt_prof_leak_error)
+CTL_EXTERN(opt_prof_pid_namespace)
+CTL_EXTERN(opt_prof_prefix)
+CTL_EXTERN(opt_prof_recent_alloc_max)
+CTL_EXTERN(opt_prof_stats)
+CTL_EXTERN(opt_prof_sys_thread_name)
+CTL_EXTERN(opt_prof_thread_active_init)
+CTL_EXTERN(opt_prof_time_res)
+CTL_EXTERN(opt_retain)
+CTL_EXTERN(opt_stats_interval)
+CTL_EXTERN(opt_stats_interval_opts)
+CTL_EXTERN(opt_stats_print)
+CTL_EXTERN(opt_stats_print_opts)
+CTL_EXTERN(opt_tcache)
+CTL_EXTERN(opt_tcache_gc_delay_bytes)
+CTL_EXTERN(opt_tcache_gc_incr_bytes)
+CTL_EXTERN(opt_tcache_max)
+CTL_EXTERN(opt_tcache_nslots_large)
+CTL_EXTERN(opt_tcache_nslots_small_max)
+CTL_EXTERN(opt_tcache_nslots_small_min)
+CTL_EXTERN(opt_thp)
+CTL_EXTERN(opt_trust_madvise)
+CTL_EXTERN(opt_utrace)
+CTL_EXTERN(opt_xmalloc)
+CTL_EXTERN(opt_zero)
+CTL_EXTERN(opt_zero_realloc)
+CTL_EXTERN(prof_active)
+CTL_EXTERN(prof_dump)
+CTL_EXTERN(prof_gdump)
+CTL_EXTERN(prof_interval)
+CTL_EXTERN(prof_log_start)
+CTL_EXTERN(prof_log_stop)
+CTL_EXTERN(prof_prefix)
+CTL_EXTERN(prof_reset)
+CTL_EXTERN(prof_stats_bins_i_accum)
+CTL_EXTERN(prof_stats_bins_i_live)
+CTL_EXTERN(prof_stats_lextents_i_accum)
+CTL_EXTERN(prof_stats_lextents_i_live)
+CTL_EXTERN(prof_thread_active_init)
+CTL_EXTERN(stats_active)
+CTL_EXTERN(stats_allocated)
+CTL_EXTERN(stats_arenas_i_abandoned_vm)
+CTL_EXTERN(stats_arenas_i_base)
+CTL_EXTERN(stats_arenas_i_bins_j_curregs)
+CTL_EXTERN(stats_arenas_i_bins_j_curslabs)
+CTL_EXTERN(stats_arenas_i_bins_j_ndalloc)
+CTL_EXTERN(stats_arenas_i_bins_j_nfills)
+CTL_EXTERN(stats_arenas_i_bins_j_nflushes)
+CTL_EXTERN(stats_arenas_i_bins_j_nmalloc)
+CTL_EXTERN(stats_arenas_i_bins_j_nonfull_slabs)
+CTL_EXTERN(stats_arenas_i_bins_j_nrequests)
+CTL_EXTERN(stats_arenas_i_bins_j_nreslabs)
+CTL_EXTERN(stats_arenas_i_bins_j_nslabs)
+CTL_EXTERN(stats_arenas_i_dirty_decay_ms)
+CTL_EXTERN(stats_arenas_i_dirty_nmadvise)
+CTL_EXTERN(stats_arenas_i_dirty_npurge)
+CTL_EXTERN(stats_arenas_i_dirty_purged)
+CTL_EXTERN(stats_arenas_i_dss)
+CTL_EXTERN(stats_arenas_i_extent_avail)
+CTL_EXTERN(stats_arenas_i_extents_j_dirty_bytes)
+CTL_EXTERN(stats_arenas_i_extents_j_muzzy_bytes)
+CTL_EXTERN(stats_arenas_i_extents_j_ndirty)
+CTL_EXTERN(stats_arenas_i_extents_j_nmuzzy)
+CTL_EXTERN(stats_arenas_i_extents_j_npinned)
+CTL_EXTERN(stats_arenas_i_extents_j_nretained)
+CTL_EXTERN(stats_arenas_i_extents_j_pinned_bytes)
+CTL_EXTERN(stats_arenas_i_extents_j_retained_bytes)
+CTL_EXTERN(stats_arenas_i_hpa_sec_bytes)
+CTL_EXTERN(stats_arenas_i_hpa_sec_dalloc_flush)
+CTL_EXTERN(stats_arenas_i_hpa_sec_dalloc_noflush)
+CTL_EXTERN(stats_arenas_i_hpa_sec_hits)
+CTL_EXTERN(stats_arenas_i_hpa_sec_misses)
+CTL_EXTERN(stats_arenas_i_hpa_sec_overfills)
+CTL_EXTERN(stats_arenas_i_hpa_shard_alloc_j_extents)
+CTL_EXTERN(stats_arenas_i_hpa_shard_alloc_j_extents_per_ps)
+CTL_EXTERN(stats_arenas_i_hpa_shard_alloc_j_max_extents)
+CTL_EXTERN(stats_arenas_i_hpa_shard_alloc_j_min_extents)
+CTL_EXTERN(stats_arenas_i_hpa_shard_alloc_j_pages_per_ps)
+CTL_EXTERN(stats_arenas_i_hpa_shard_alloc_j_ps)
+CTL_EXTERN(stats_arenas_i_hpa_shard_alloc_j_total_elapsed_ns_per_ps)
+CTL_EXTERN(stats_arenas_i_hpa_shard_empty_slabs_nactive_huge)
+CTL_EXTERN(stats_arenas_i_hpa_shard_empty_slabs_nactive_nonhuge)
+CTL_EXTERN(stats_arenas_i_hpa_shard_empty_slabs_ndirty_huge)
+CTL_EXTERN(stats_arenas_i_hpa_shard_empty_slabs_ndirty_nonhuge)
+CTL_EXTERN(stats_arenas_i_hpa_shard_empty_slabs_npageslabs_huge)
+CTL_EXTERN(stats_arenas_i_hpa_shard_empty_slabs_npageslabs_nonhuge)
+CTL_EXTERN(stats_arenas_i_hpa_shard_full_slabs_nactive_huge)
+CTL_EXTERN(stats_arenas_i_hpa_shard_full_slabs_nactive_nonhuge)
+CTL_EXTERN(stats_arenas_i_hpa_shard_full_slabs_ndirty_huge)
+CTL_EXTERN(stats_arenas_i_hpa_shard_full_slabs_ndirty_nonhuge)
+CTL_EXTERN(stats_arenas_i_hpa_shard_full_slabs_npageslabs_huge)
+CTL_EXTERN(stats_arenas_i_hpa_shard_full_slabs_npageslabs_nonhuge)
+CTL_EXTERN(stats_arenas_i_hpa_shard_nactive)
+CTL_EXTERN(stats_arenas_i_hpa_shard_ndehugifies)
+CTL_EXTERN(stats_arenas_i_hpa_shard_ndirty)
+CTL_EXTERN(stats_arenas_i_hpa_shard_nhugifies)
+CTL_EXTERN(stats_arenas_i_hpa_shard_nhugify_failures)
+CTL_EXTERN(stats_arenas_i_hpa_shard_nonfull_slabs_j_nactive_huge)
+CTL_EXTERN(stats_arenas_i_hpa_shard_nonfull_slabs_j_nactive_nonhuge)
+CTL_EXTERN(stats_arenas_i_hpa_shard_nonfull_slabs_j_ndirty_huge)
+CTL_EXTERN(stats_arenas_i_hpa_shard_nonfull_slabs_j_ndirty_nonhuge)
+CTL_EXTERN(stats_arenas_i_hpa_shard_nonfull_slabs_j_npageslabs_huge)
+CTL_EXTERN(stats_arenas_i_hpa_shard_nonfull_slabs_j_npageslabs_nonhuge)
+CTL_EXTERN(stats_arenas_i_hpa_shard_npageslabs)
+CTL_EXTERN(stats_arenas_i_hpa_shard_npurge_passes)
+CTL_EXTERN(stats_arenas_i_hpa_shard_npurges)
+CTL_EXTERN(stats_arenas_i_hpa_shard_slabs_nactive_huge)
+CTL_EXTERN(stats_arenas_i_hpa_shard_slabs_nactive_nonhuge)
+CTL_EXTERN(stats_arenas_i_hpa_shard_slabs_ndirty_huge)
+CTL_EXTERN(stats_arenas_i_hpa_shard_slabs_ndirty_nonhuge)
+CTL_EXTERN(stats_arenas_i_hpa_shard_slabs_npageslabs_huge)
+CTL_EXTERN(stats_arenas_i_hpa_shard_slabs_npageslabs_nonhuge)
+CTL_EXTERN(stats_arenas_i_internal)
+CTL_EXTERN(stats_arenas_i_large_allocated)
+CTL_EXTERN(stats_arenas_i_large_ndalloc)
+CTL_EXTERN(stats_arenas_i_large_nfills)
+CTL_EXTERN(stats_arenas_i_large_nflushes)
+CTL_EXTERN(stats_arenas_i_large_nmalloc)
+CTL_EXTERN(stats_arenas_i_large_nrequests)
+CTL_EXTERN(stats_arenas_i_lextents_j_curlextents)
+CTL_EXTERN(stats_arenas_i_lextents_j_ndalloc)
+CTL_EXTERN(stats_arenas_i_lextents_j_nmalloc)
+CTL_EXTERN(stats_arenas_i_lextents_j_nrequests)
+CTL_EXTERN(stats_arenas_i_mapped)
+CTL_EXTERN(stats_arenas_i_metadata_edata)
+CTL_EXTERN(stats_arenas_i_metadata_rtree)
+CTL_EXTERN(stats_arenas_i_metadata_thp)
+CTL_EXTERN(stats_arenas_i_muzzy_decay_ms)
+CTL_EXTERN(stats_arenas_i_muzzy_nmadvise)
+CTL_EXTERN(stats_arenas_i_muzzy_npurge)
+CTL_EXTERN(stats_arenas_i_muzzy_purged)
+CTL_EXTERN(stats_arenas_i_nthreads)
+CTL_EXTERN(stats_arenas_i_pactive)
+CTL_EXTERN(stats_arenas_i_pdirty)
+CTL_EXTERN(stats_arenas_i_pinned)
+CTL_EXTERN(stats_arenas_i_pmuzzy)
+CTL_EXTERN(stats_arenas_i_resident)
+CTL_EXTERN(stats_arenas_i_retained)
+CTL_EXTERN(stats_arenas_i_small_allocated)
+CTL_EXTERN(stats_arenas_i_small_ndalloc)
+CTL_EXTERN(stats_arenas_i_small_nfills)
+CTL_EXTERN(stats_arenas_i_small_nflushes)
+CTL_EXTERN(stats_arenas_i_small_nmalloc)
+CTL_EXTERN(stats_arenas_i_small_nrequests)
+CTL_EXTERN(stats_arenas_i_tcache_bytes)
+CTL_EXTERN(stats_arenas_i_tcache_stashed_bytes)
+CTL_EXTERN(stats_arenas_i_uptime)
+CTL_EXTERN(stats_background_thread_num_runs)
+CTL_EXTERN(stats_background_thread_num_threads)
+CTL_EXTERN(stats_background_thread_run_interval)
+CTL_EXTERN(stats_mapped)
+CTL_EXTERN(stats_metadata)
+CTL_EXTERN(stats_metadata_edata)
+CTL_EXTERN(stats_metadata_rtree)
+CTL_EXTERN(stats_metadata_thp)
+CTL_EXTERN(stats_mutexes_reset)
+CTL_EXTERN(stats_pinned)
+CTL_EXTERN(stats_resident)
+CTL_EXTERN(stats_retained)
+CTL_EXTERN(stats_zero_reallocs)
+
+#define MUTEX_STATS_CTL_EXTERN_GEN(n)                                         \
+	CTL_EXTERN(stats_##n##_num_ops)                                       \
+	CTL_EXTERN(stats_##n##_num_wait)                                      \
+	CTL_EXTERN(stats_##n##_num_spin_acq)                                  \
+	CTL_EXTERN(stats_##n##_num_owner_switch)                              \
+	CTL_EXTERN(stats_##n##_total_wait_time)                               \
+	CTL_EXTERN(stats_##n##_max_wait_time)                                 \
+	CTL_EXTERN(stats_##n##_max_num_thds)
+
+/* Global mutexes. */
+#define OP(mtx) MUTEX_STATS_CTL_EXTERN_GEN(mutexes_##mtx)
+MUTEX_PROF_GLOBAL_MUTEXES
+#undef OP
+
+/* Per arena mutexes. */
+#define OP(mtx) MUTEX_STATS_CTL_EXTERN_GEN(arenas_i_mutexes_##mtx)
+MUTEX_PROF_ARENA_MUTEXES
+#undef OP
+
+/* Arena bin mutexes. */
+MUTEX_STATS_CTL_EXTERN_GEN(arenas_i_bins_j_mutex)
+#undef MUTEX_STATS_CTL_EXTERN_GEN
+
+CTL_EXTERN(tcache_create)
+CTL_EXTERN(tcache_destroy)
+CTL_EXTERN(tcache_flush)
+CTL_EXTERN(thread_allocated)
+CTL_EXTERN(thread_allocatedp)
+CTL_EXTERN(thread_arena)
+CTL_EXTERN(thread_deallocated)
+CTL_EXTERN(thread_deallocatedp)
+CTL_EXTERN(thread_idle)
+CTL_EXTERN(thread_peak_read)
+CTL_EXTERN(thread_peak_reset)
+CTL_EXTERN(thread_prof_active)
+CTL_EXTERN(thread_prof_name)
+CTL_EXTERN(thread_tcache_enabled)
+CTL_EXTERN(thread_tcache_flush)
+CTL_EXTERN(thread_tcache_max)
+CTL_EXTERN(thread_tcache_ncached_max_read_sizeclass)
+CTL_EXTERN(thread_tcache_ncached_max_write)
+
+#undef CTL_EXTERN
+
+#endif /* JEMALLOC_INTERNAL_CTL_EXTERNS_H */
